requirements.txt
pyspark==4.0.1
pandas
pytest
pyyaml
groq
